// src/types/acara.ts
export type Product = {
  pembicara: string;
  photo_url: string;
  name: string;
  description: string;
  created_at: string;
  harga: string;
  jabatan_pembicara: string;
  id: string;
  tanggal: Date;
  category: string;
  is_free: boolean;
  is_public: boolean;
  updated_at: string;
  slot: string;
  tempat: string;
  thumbnailUrl: string;
};